modularize cast and reshape
