+++ tfjs-node/src/kernels/Reshape.ts
@@ -1,0 +1,40 @@
+/**
+ * @license
+ * Copyright 2020 Google LLC. All Rights Reserved.
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * =============================================================================
+ */
+
+import {KernelConfig, Reshape, ReshapeAttrs, ReshapeInputs, tensor1d} from '@tensorflow/tfjs';
+
+import {createTensorsTypeOpAttr, NodeJSKernelBackend} from '../nodejs_kernel_backend';
+
+export const reshapeConfig: KernelConfig = {
+  kernelName: Reshape,
+  backendName: 'tensorflow',
+  kernelFunc: (args) => {
+    const {x} = args.inputs as ReshapeInputs;
+    const backend = args.backend as NodeJSKernelBackend;
+    const {shape} = args.attrs as {} as ReshapeAttrs;
+
+    const shapeTensor = tensor1d(shape, 'int32');
+
+    const opAttrs = [
+      createTensorsTypeOpAttr('T', x.dtype),
+      createTensorsTypeOpAttr('Tshape', shapeTensor.dtype)
+    ];
+    const res = backend.executeSingleOutput(Reshape, opAttrs, [x, shapeTensor]);
+    shapeTensor.dispose();
+    return res;
+  }
+};
