+++ tfjs-node/src/register_all_kernels.ts
@@ -40,7 +40,9 @@ import {avgPool3DGradConfig} from './kernels/AvgPool3DGrad';
 import {avgPoolGradConfig} from './kernels/AvgPoolGrad';
 import {batchMatMulConfig} from './kernels/BatchMatMul';
 import {batchToSpaceNDConfig} from './kernels/BatchToSpaceND';
+import {castConfig} from './kernels/Cast';
 import {ceilConfig} from './kernels/Ceil';
+import {clipByValueConfig} from './kernels/ClipByValue';
 import {concatConfig} from './kernels/Concat';
 import {conv2DConfig} from './kernels/Conv2D';
 import {conv2DBackpropFilterConfig} from './kernels/Conv2DBackpropFilter';
@@ -113,6 +115,7 @@ import {realDivConfig} from './kernels/RealDiv';
 import {reciprocalConfig} from './kernels/Reciprocal';
 import {reluConfig} from './kernels/Relu';
 import {relu6Config} from './kernels/Relu6';
+import {reshapeConfig} from './kernels/Reshape';
 import {resizeBilinearConfig} from './kernels/ResizeBilinear';
 import {resizeBilinearGradConfig} from './kernels/ResizeBilinearGrad';
 import {resizeNearestNeighborConfig} from './kernels/ResizeNearestNeighbor';
@@ -148,140 +151,141 @@ import {transposeConfig} from './kernels/Transpose';
 import {unpackConfig} from './kernels/Unpack';
 import {unsortedSegmentSumConfig} from './kernels/UnsortedSegmentSum';
 import {zerosLikeConfig} from './kernels/ZerosLike';
-import { clipByValueConfig } from "./kernels/ClipByValue";
 
 // List all kernel configs here
 const kernelConfigs: KernelConfig[] = [
-LRNConfig,
-LRNGradConfig,
-_fusedMatMulConfig,
-absConfig,
-acosConfig,
-acoshConfig,
-addConfig,
-addNConfig,
-allConfig,
-anyConfig,
-argMaxConfig,
-argMinConfig,
-asinConfig,
-asinhConfig,
-atan2Config,
-atanConfig,
-atanhConfig,
-avgPool3DConfig,
-avgPool3DGradConfig,
-avgPoolConfig,
-avgPoolGradConfig,
-batchMatMulConfig,
-batchToSpaceNDConfig,
-ceilConfig,
-clipByValueConfig,
-concatConfig,
-conv2DBackpropFilterConfig,
-conv2DBackpropInputConfig,
-conv2DConfig,
-cosConfig,
-coshConfig,
-cropAndResizeConfig,
-cumsumConfig,
-depthToSpaceConfig,
-depthwiseConv2dNativeBackpropFilterConfig,
-depthwiseConv2dNativeBackpropInputConfig,
-depthwiseConv2dNativeConfig,
-diagConfig,
-dilation2dBackpropFilterConfig,
-dilation2dBackpropInputConfig,
-dilation2dConfig,
-eluConfig,
-eluGradConfig,
-equalConfig,
-erfConfig,
-expConfig,
-expm1Config,
-fillConfig,
-floorConfig,
-floorDivConfig,
-fusedBatchNormConfig,
-fusedConv2DConfig,
-fusedDepthwiseConv2DConfig,
-gatherNdConfig,
-gatherV2Config,
-greaterConfig,
-greaterEqualConfig,
-isFiniteConfig,
-isInfConfig,
-isNanConfig,
-lessConfig,
-lessEqualConfig,
-linSpaceConfig,
-log1pConfig,
-logConfig,
-logicalAndConfig,
-logicalNotConfig,
-logicalOrConfig,
-maxConfig,
-maxPool3DConfig,
-maxPool3DGradConfig,
-maxPoolConfig,
-maxPoolGradConfig,
-maximumConfig,
-minConfig,
-minimumConfig,
-mirrorPadConfig,
-modConfig,
-multinomialConfig,
-multiplyConfig,
-negConfig,
-nonMaxSuppressionV3Config,
-nonMaxSuppressionV4Config,
-nonMaxSuppressionV5Config,
-notEqualConfig,
-oneHotConfig,
-onesLikeConfig,
-padV2Config,
-powConfig,
-preluConfig,
-prodConfig,
-realDivConfig,
-reciprocalConfig,
-relu6Config,
-reluConfig,
-resizeBilinearConfig,
-resizeBilinearGradConfig,
-resizeNearestNeighborConfig,
-resizeNearestNeighborGradConfig,
-reverseConfig,
-roundConfig,
-rsqrtConfig,
-scatterNdConfig,
-selectConfig,
-seluConfig,
-sigmoidConfig,
-signConfig,
-sinConfig,
-sinhConfig,
-sliceConfig,
-softmaxConfig,
-softplusConfig,
-spaceToBatchNDConfig,
-sparseToDenseConfig,
-splitVConfig,
-sqrtConfig,
-squareConfig,
-squaredDifferenceConfig,
-stepConfig,
-stridedSliceConfig,
-subConfig,
-sumConfig,
-tanConfig,
-tanhConfig,
-tileConfig,
-topKConfig,
-transposeConfig,
-unpackConfig,
-unsortedSegmentSumConfig,
-zerosLikeConfig
+  LRNConfig,
+  LRNGradConfig,
+  _fusedMatMulConfig,
+  absConfig,
+  acosConfig,
+  acoshConfig,
+  addConfig,
+  addNConfig,
+  allConfig,
+  anyConfig,
+  argMaxConfig,
+  argMinConfig,
+  asinConfig,
+  asinhConfig,
+  atan2Config,
+  atanConfig,
+  atanhConfig,
+  avgPool3DConfig,
+  avgPool3DGradConfig,
+  avgPoolConfig,
+  avgPoolGradConfig,
+  batchMatMulConfig,
+  batchToSpaceNDConfig,
+  castConfig,
+  ceilConfig,
+  clipByValueConfig,
+  concatConfig,
+  conv2DBackpropFilterConfig,
+  conv2DBackpropInputConfig,
+  conv2DConfig,
+  cosConfig,
+  coshConfig,
+  cropAndResizeConfig,
+  cumsumConfig,
+  depthToSpaceConfig,
+  depthwiseConv2dNativeBackpropFilterConfig,
+  depthwiseConv2dNativeBackpropInputConfig,
+  depthwiseConv2dNativeConfig,
+  diagConfig,
+  dilation2dBackpropFilterConfig,
+  dilation2dBackpropInputConfig,
+  dilation2dConfig,
+  eluConfig,
+  eluGradConfig,
+  equalConfig,
+  erfConfig,
+  expConfig,
+  expm1Config,
+  fillConfig,
+  floorConfig,
+  floorDivConfig,
+  fusedBatchNormConfig,
+  fusedConv2DConfig,
+  fusedDepthwiseConv2DConfig,
+  gatherNdConfig,
+  gatherV2Config,
+  greaterConfig,
+  greaterEqualConfig,
+  isFiniteConfig,
+  isInfConfig,
+  isNanConfig,
+  lessConfig,
+  lessEqualConfig,
+  linSpaceConfig,
+  log1pConfig,
+  logConfig,
+  logicalAndConfig,
+  logicalNotConfig,
+  logicalOrConfig,
+  maxConfig,
+  maxPool3DConfig,
+  maxPool3DGradConfig,
+  maxPoolConfig,
+  maxPoolGradConfig,
+  maximumConfig,
+  minConfig,
+  minimumConfig,
+  mirrorPadConfig,
+  modConfig,
+  multinomialConfig,
+  multiplyConfig,
+  negConfig,
+  nonMaxSuppressionV3Config,
+  nonMaxSuppressionV4Config,
+  nonMaxSuppressionV5Config,
+  notEqualConfig,
+  oneHotConfig,
+  onesLikeConfig,
+  padV2Config,
+  powConfig,
+  preluConfig,
+  prodConfig,
+  realDivConfig,
+  reciprocalConfig,
+  relu6Config,
+  reluConfig,
+  reshapeConfig,
+  resizeBilinearConfig,
+  resizeBilinearGradConfig,
+  resizeNearestNeighborConfig,
+  resizeNearestNeighborGradConfig,
+  reverseConfig,
+  roundConfig,
+  rsqrtConfig,
+  scatterNdConfig,
+  selectConfig,
+  seluConfig,
+  sigmoidConfig,
+  signConfig,
+  sinConfig,
+  sinhConfig,
+  sliceConfig,
+  softmaxConfig,
+  softplusConfig,
+  spaceToBatchNDConfig,
+  sparseToDenseConfig,
+  splitVConfig,
+  sqrtConfig,
+  squareConfig,
+  squaredDifferenceConfig,
+  stepConfig,
+  stridedSliceConfig,
+  subConfig,
+  sumConfig,
+  tanConfig,
+  tanhConfig,
+  tileConfig,
+  topKConfig,
+  transposeConfig,
+  unpackConfig,
+  unsortedSegmentSumConfig,
+  zerosLikeConfig
 ];
 
 for (const kernelConfig of kernelConfigs) {
